[MOVE] Internal classes to internal package
